Scrawls (#23)
* Turnin this thing into an astro blog

Turned the whole thing into an astro blog.

Listed blog posts on homepage
Got the metShow component working, it pulls data from the met endpoint.

* Loading Images onto Galleries

The framework for auto image galleries is done, can't wait to style it.

* Fonts and gallery implemented

Custom Fonts have been brought in and a simple grid has been implemented
for the galleries

* Fixes fonts not being displayed properly

moving over to @ font-face declarations instead of direct link to fonts
for better browser support.

* ignores Ds_store

why was this not being ignored???

* This commit adds a new notes collection

A shorter form note will be more prevalent on the blog, so it makes
sense to add an individual collection for it instead of filtering them
by tags or other means.

* This one gives mee a blank slate to work with on astro

new beginings, this is just getting thigs started and probably will be
in dev for a hot second.

* This site is being redesigned

This commit ads a simple disclaimer that tells everyone that it looks
awful because its being redesigned

---------

Co-authored-by: Rodrigo Ibarra <rodrigoibarra@192.168.8.103>